ts: Move iterator tests to separate file
Release note: None
